preparation for moving lyra out of Celesta
